Prevent integer overflow - Bug #19895
git-svn-id: file:///svn/phpbb/branches/phpBB-3_0_0@8876 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/includes/acp/acp_main.php
@@ -166,7 +166,7 @@ class acp_main
 							FROM ' . ATTACHMENTS_TABLE . '
 							WHERE is_orphan = 0';
 						$result = $db->sql_query($sql);
-						set_config('upload_dir_size', (int) $db->sql_fetchfield('stat'), true);
+						set_config('upload_dir_size', (float) $db->sql_fetchfield('stat'), true);
 						$db->sql_freeresult($result);
 
 						if (!function_exists('update_last_username'))
